Fix seq handling: enforce integers, add renumber_scenes
Seq values must always be integers. When scenes are inserted, reordered,
split, or merged, the scenes skill now renumbers seq sequentially from 1.

- Add renumber_scenes() to csv.sh — sorts by current seq, renumbers 1..N
- Revert prompts.py float() back to int() — seq is not fractional
- Keep revision.py int() — same rationale
- Cleanup script detects gaps/non-integers as SEQ_NEEDS_RENUMBER
- Scenes skill documents renumber-after-every-structural-change rule
- 4 new tests for renumber_scenes

Fixes #49

+++ scripts/lib/python/storyforge/prompts.py
@@ -252,7 +252,7 @@ def get_previous_scene(scene_id: str, project_dir: str) -> str:
     for row in rows:
         if len(row) > max(id_idx, seq_idx):
             try:
-                seq_val = float(row[seq_idx]) if row[seq_idx] else 0
+                seq_val = int(row[seq_idx]) if row[seq_idx] else 0
             except ValueError:
                 seq_val = 0
             pairs.append((row[id_idx], seq_val))
